overlord: wait for snapd restart after requesting by undo of 'link-snap' (#14850)
wait for snapd restart after requesting by undo of 'link-snap'. otherwise we can end up using tools from the snapd we are reverting away from, causing issues or generation of wrong apparmor profiles

* overlord/ifacestate: correct the comment

* tests/core/snapd-refresh-undo: couple of linting issues

* tests,overlord: rectify comments, add a TODO for the restart request in undoUnlinkCurrentSnap, fix two unit tests

* overlord/devicestate: fix double locking in unit test

* overlord/snapstate: add sequence file check

* overlord/snapstate: add TODO for undo path relating to snap-failure

* tests/core/snapd-refresh-undo: revert snapd in end of test to fix restore

* overlord/devicestate: make unit test more robust

* overlord/snapstate,tests: use correct terminology for snapd restarts, and fix the description of the spread test

* wrappers: more logging

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* wrappers: snapd version specific quirk for < 2.62 versions

Introduce a version specific quirk which works around a problem present
in snapd in versions < 2.62. Prior to 2.622 the snapd.apparmor.service
did not carry a do-not-start tag. This could lead to a problem where
when executing the undo path of snapd, the snapd.apparmor.service is
restarted after unlinking the current version of snapd. This could cause
the snapd.apparmor to use either a potentially much older parser from
the host one from the previous version of snapd, where in either case,
the parser may not support features present in the profiles generated by
current version of snapd.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>
Co-authored-by: Maciej Borzecki <maciej.borzecki@canonical.com>