osutil/user: look up getent executable in known host directories (#14792)
* osutil/user: search for getent executable in known host directories

It is possible that `snap run` may be called with a PATH set such that
it does not include a directory containing getent. To workaround this,
try to search for getent in a number of known locations on the host.

Fixes: https://bugs.launchpad.net/snapd/+bug/2090938

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* osutil/user: allow overriding getent search path

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* osutil: update user lookup tests for possible build with snapdusergo tag

The snapdusergo tag disables use of cgo/Go based user.Lookup() (and
similar) code, and instead provides alternative implementations that use
getent under the hood. Adjust the tests to account for this.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>