o/c/configcore,o/ifacestate: check for presence of handler-service on startup (#14843)
* o/c/configcore,o/ifacestate: check for presence of handler-service on startup

During snapd startup, if the "apparmor-prompting" flag is enabled but
there is no snap with an interfaces-requests-control connection and an
app app declared as a "handler-service", then record a warning that
prompts will be auto-denied until a prompting client is installed.

Such a situation could arise if the user uninstalled the prompting
client snap without disabling the "apparmor-prompting" feature flag, or
if they disconnected the plug for the interfaces-requests-control
interface.

There is a check for the presence of a handler-service client when one
attempts to enable the "apparmor-prompting" feature flag in the first
place: when enabling the flag, if there is no handler-service app
present, or it cannot be successfully started, then the flag is left
disabled. Before this commit, however, there were no subsequent checks
for the presence of a handler-service, so the user would have no
indication of why snaps may be blocking on syscalls or having syscalls
denied. Now, we make this check again at snapd startup, but rather than
disabling prompting, we instead simply record a warning to the user.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/ifacestate: simplify mocking of function to check handler service presence

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* o/ifacestate: improve comment around handler service presence error

Co-authored-by: Miguel Pires <miguelpires94@gmail.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Miguel Pires <miguelpires94@gmail.com>