core-initrd: do not mount /lib/{firmware,modules} in run mode
Do not mount /lib/{firmware,modules} in run mode as now snap-bootstrap
takes care of doing the right mounts in that case for UC24+, mounting
the drivers tree on these folders. We still need to do these mounts
for the other modes, as there is no drivers tree in that case.