many: rework the way we allocate counter handles
We used 4 handles from 0x01880001-0x01880004 for PCR policies. We
always used a pair and would switch to the other pair after factory
reset.

No instead we look for free handle in range 0x01880005-0x0188000F. We
also use only one handle per primary key.

When we factory reset, we should remove the handles correctly since
the old primary key for the save partition should have been the same
as the data partition that was removed, thus the handles should be the
same.

We also now correctly populate the FDE state with the correct handles.