snap-bootstrap: load partition information from a manifest file for cloudimg-rootfs mode (#14713)
* c/snap-bootstrap: refactor CVM mode

This is a preparatory commit which modifies the existing CVM mode
to support mounting multiple overlaid partitions for the rootfs.

* c/snap-bootstrap: support mount overlays in CVM mode

In order to provide integrity protection for the root fs, the root fs will
be mounted read-only and will be accompanied by dm-verity data from a new
auto-discovered partition. To support writable state for the ephemeral CVM
scenario, an overlayfs mount is created with an upper layer residing on tmpfs.
A writable layer residing on disk can be also supported but it is currently
disabled.

The filesystem layout in the initramfs will look like:

. /run/mnt
├── ...
├── data                     (where the final merged overlay will be mounted)
├── cloudimg-rootfs          (read-only mount of the root fs, folder name is the fs label)
├── ..                       (future WIP, other partition mounts as separate mounts)
└── cloudimg-rootfs-writable (tmpfs mount for the writable upper layer)
    ├── work
    └── upper

An example final overlay mount command will look like:

mount -t overlay overlay -olowerdir=/run/mnt/cloudimg-rootfs,\
upperdir=/run/mnt/cloudimg-rootfs-writable/upper,\
workdir=/run/mnt/cloudimg-rootfs-writable/work \
/run/mnt/data

* c/snap-bootstrap: use partition label to auto-discover the rootfs dm-verity partition instead

* c/snap-bootstrap: refactor manifest parsing code for CVM mode

* c/snap-bootstrap: rename FsLabel field to GptLabel

v1 CVM images use the filesystem label to auto-discover an encrypted
rootfs partition based on whether it has the '-enc' suffix. The new
images use the GPT label to auto-discover the verity partition. In
order to keep backwards compatibility, the customization tool will use
set both labels to the same one.

This commit clarifies that the field in the new structs actually contain
GPT label information.

* c/snap-bootstrap: change path name of writable tmpfs partition

* c/snap-bootstrap: make comment more generic

* c/snap-bootstrap: limit upper overlay folder creation to the tmpfs case

creating the upper and work folders used for setting up the overlayfs
mount, was initially handled by snap-bootstrap whether it set up a tmpfs
or detected a writable partition on disk but in order to support the
writable partition case, it was later discovered that these folders might
need to be created during the customization step and might already exist.

snap-bootstrap will now create those folders only if it has detected
no writable partition in the manifest and has auto-created a tmpfs-based
writable partition.

* c/snap-bootstrap: revert comment for secbootProvisionForCVM

trying to boot an image on a system without a TPM present was failing
due to not-finding a TPM device. The source of the error was not the
provision step as secbotProvisionForCVM already ignores errors (returns
nil) if a TPM device is not found or a template file for provision the
TPM is not found.

* c/snap-bootstrap: fix test for CVM mode

CVM mode used to mount the rootfs partition with the Ephemeral flag as
cloud images contain a fixed fstab entry.

* c/snap-bootstrap: remove redundant disk check

This is a legacy check that is not actually needed now as
secbootUnlockVolumeUsingSealedKeyIfEncrypted will only find partitions
residing on the disk that the ESP is from.

* c/snap-bootstrap: add test for CVM ephemeral mode with manifest

* c/snap-bootstrap: make cvm structures private

* c/snap-bootstrap: create a manifest file in test instead of mocking the read

* c/snap-bootstrap: simplify manifest and add extra tests

* c/snap-bootstrap: minor comment changes in test cases for cvm

* c/snap-bootstrap: address comment wrt to manifest handling in cvm mode

* c/snap-bootstrap: add tests for createOverlayDirs

* cmd/snap-bootstrap: add TODO notice for manifest measurement

* c/snap-bootstrap: address comments for cvm mode

* cmd/snap-bootstrap: fix comment for inclusive language

* cmd/snap-bootstrap: remove unused manifestError