interfaces: disable udev backend when inside a container
Devices are not namespace-aware in Linux, thus udev inside containers
does not really do anything sensible and is set to not execute if /sys
is not mounted as a writable file system.

This solves the issue where snapd is failing to install snaps in LXD,
demanding workarounds like installing the snap twice, hoping that
once-generated files are intact and not trigger a udev rule reload.

Fixes: https://bugs.launchpad.net/snapd/+bug/1712808
Fixes: https://bugs.launchpad.net/snapd/+bug/1865503
Jira: https://warthogs.atlassian.net/browse/SNAPDENG-24757

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>