cmd/snap-bootstrap: mount drivers tree if present (#14522)
* cmd/snap-bootstrap: mount drivers tree if present

Mount drivers tree if present and do not mount anymore the kernel in
/run/mnt/kernel in that case. Mount also the kernel snap in the usual
location under /snap so symbolic links in /lib/{firmware,modules}
point to something that exists.

* core-initrd: do not mount /lib/{firmware,modules} in run mode

Do not mount /lib/{firmware,modules} in run mode as now snap-bootstrap
takes care of doing the right mounts in that case for UC24+, mounting
the drivers tree on these folders. We still need to do these mounts
for the other modes, as there is no drivers tree in that case.

* tests/basic20plus: adapt to changes in UC24+

We are not mounting the kernel snap anymore in UC24+.

* tests/nested/manual/muinstaller-real: do not customize extra-paths

This is no longer necessary as now we build the initramfs using files
from the snapd repository.