o/snapstate: put MaybeReboot edges on link-snap in AddLinkNewBaseOrKernel and LinkNewBaseOrKernel
Sometimes during a remodel, we need to create link-snap tasks for snaps
that are already installed, but not yet considered boot participants.
These tasks were missing reboot boundaries, which caused the tasks to be
marked as done before the reboot completed.

tests/nested/manual/remodel-offline exhibits this behavior, where core22
already is installed, but not the model's base snap. The link-snap task
that was created during the remodel was being marked as done before the
reboot that the task triggers completed.

With this change, the entire remodel change will not be marked as done
until that link-snap task triggers is completed