daemon,i/prompting: replace PermissionsListEmptyError with PermissionsEmptyError
Previously, both prompts and rules used lists of permissions, but now
that rules use a map from permission to outcome and lifespan, language
around permission lists is no longer accurate. Simply remove the "list"
from the error name and error message, and it can continue to be used in
both situations.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>