interfaces: add configfiles backend
The configfiles backend ensures the state of configuration files in
the filesystem. Interfaces using it must provide a list of globs with
the files that are under their control.

For the moment, only classic systems are supported (we write the
configuration files in the rootfs). In the future, doing similar
changes in the mount namespace of the application might be
implemented.