many: abbreviate version from git by three bytes
The snap ecosystem enforces a length limit on snap versions and the
version of snapd, especially non-tagged versions that include git hashes
and the FIPS marker, are simply too long.

Abbreviate the longest variant of the version by three bytes by dropping
the word "git". This in itself should not impact sorting as the plus
byte following git is retained.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>