i/prompting: do not error when converting empty abstract perms to AA perms (#14975)
It's reasonable that one might wish to convert an empty list of abstract
permissions to AppArmor permissions, such as when one is constructing an
allowed permissions set where all permissions are denied.

It was incorrect to treat this as an error, and furthermore, the only
current caller of `AbstractPermissionsToAppArmorPermissions` which
actually uses the error value is in `requestprompts`, which logs a
logger notice on an error. So we were getting false-positive errors in
the journalctl log. This commit fixes that issue.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>