tests: skip confdb-cross-config on debian11
This test sometimes fails due to snap-confine issues when running after
a specific set of tests. I haven't been able to debug it yet but it
doesn't happen if run by itself, on another system or even with a
different set of sets running before, so it's likely not a confdb issue.
I'm disabling this for debian11 while I debug it further so it doesn't
impact other work.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>