store: add functions to download snap icons
These new functions, `downloadIconImpl` and `DownloadIcon`, mirror their
non-icon counterparts `downloadImpl` and `Download`, except that because
snap icons are small, simple assets on a fileserver, there is no need
for snap-specific headers, authentication, partial downloads, or
caching.

These will be used in the future in order to download a snap's icon
whenever the new `.snap` blob is downloaded for that snap and linked to
a snap icons directory so that the snapd API can serve snap icons from
local icon files, rather than returning an error if the snap does not
ship an icon at `meta/gui/icon.*`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>