cmd/snap-confine: fix order of container checks
When snap-confine is executing inside container it doesn't do anything
with devices but the order of checks is such, that the function
sc_get_device_cgroup_setup is called unconditionally. Since the recent
change to snapd udev backend, that function will fail to load the
snap.$SNAP_NAME.device file when executing inside a container.

Re-arrange the order so that container check happens earlier.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>