cmd, github: consistently use clang-format (#14971)
* cmd: apply clang-format to all C files

Apply clang-format to all C files in the tree.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd: stop formatting code with indent, use clang-format

We have historically avoided reformatting all code with clang-format,
with the main motivator being more complicated tracking of previous
changes. However with tight integration of git blame in various IDEs,
github UI and even Emacs, following change history is now trivial.

Additionally GNU indent itself is a source of issues.. The version which
was present included in 22.04 (2.2.12) was notoriously producing
incorrect formatting, to the point that the CI process in snapd built
its own 2.2.13 binary. On top of this, formatting with indent does not
produce stable results. One has to run indent at least twice to produce
stable output. In contrast, clang-format, at least 19.x which I use at
time this change is done, has none of those issues and has been used in
community with success.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* github: stop building indent 2.2.13

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>