many: make kernel-modules components available on initramfs installation (#14810)
* asserts,overlord: move NeedsKernelSetup to a model method

* gadget,o/devicestate: move build of kernel info for boot

to install package so it can be used by snap-bootstrap.

* seed,overlord: add new component to snap for unit testing

* cmd/snap-bootstrap: make kernel-modules components available

on first boot, when installing from the initramfs.

* tests: introduce vars to create kernel comps for nested

Introduce NESTED_COMP_KERNEL_MODULE_NAME and
NESTED_KERNEL_MODULES_COMP env vars. If present, a kernel-modules
component containing the specified kernel module will be created and
will be added when building the nested test image.

* many: create KernelBootInfo struct to be returned by BuildKernelBootInfo

* cmd/snap-bootstrap: create drivers tree mounts when installing

Create the drivers tree mounts when installing UC from the initramfs.

* cmd/snap-bootstrap: unmount properly component mounts used just

to build the drivers tree.

* tests/uc20-install-in-initrd: add checks for mount files

* o/snapstate: consider components on preseeding

Run the prepare-kernel-modules-components task when creating the
pre-seed tarball. This was not happening as pre-seeding stops the run
of snap install tasks right before running hooks, and this task is run
after them. To workaround this, add a
prepare-kernel-modules-components task that runs before hooks, only
for preseeding.

* cmd/snap-bootstrap: do not create drivers tree if preseeding

It will be part of the preseed tarball.

* many: move NeedsKernelSetup to kernel package

* many: move BuildKernelBootInfo to overlord/install package

* many: some refactoring and name changes

* tests: do not use kernel-modules components when UC < UC24