o/snapstate: download and manage snap icons
During the "download-snap" task, download the snap icon as well.

During "link-snap" and "unlink-snap", link/unlink the snap icon from the
downloaded icons pool to the the installed icons directory, managed much
the same way is snap auxinfo. This is because snap icons, like auxinfo,
are not revision-specific. Instead, they reflect the current state of
the snap listing in the snap store.

Lastly, discard the downloaded snap icon when the final revision of the
snap is discarded from disk. We want to ensure that the snap icon
remains in the pool during any situation when it's possible to install
the snap without doing another "download-snap" task, such as via a
revert.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>