core-initrd: adapt to changes now that the base mounts to /sysroot
With the snap-bootstrap changes, the base/rootfs is mounted directly
to /sysroot. With this, we do not need the symlinks created by
detect-*-sysroot.service units. However, we need to keep the static
sysroot mount unit so dependencies are still honored.

kernel-snap-generator and classic-mounts.service are not needed as for
24+ the kernel mount units are created by snap-bootstrap.

The units that need to detect if we are on core/hybrid check now for
the /run/mnt/data/system-data (if present, it is UC) instead of for a
mountpoint in /run/mnt/base.