confdb: support marking paths in schemas as ephemeral
This adds support for marking parts of a databag as ephemeral, by
allowing us to add a `"ephemeral": true` constraint to type definitions.
It also adds an `Ephemeral() bool` function to expose this. In reality,
I expect that the most natural implementation of the pruning will just
recursively prune parts of the data based on the ephemeral field without
using this function directly but I needed some way of testing the
parsing without having to also implement the pruning in this commit.
We also use Ephemeral() to forbid user-defined types from using
ephemeral so, when we support that, we can remove it from the interface
and move the functions to export_test just for white-box testing.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>