o/snapstate: remove snap icon from pool only during doDiscardSnap
The lifecycle of snap icons in the icons pool directory should mirror
that of `.snap` files, and only be cleaned up during `doDiscardSnap`, if
there are no other revisions of the snap present on the system. There is
no need for the `undoDownloadSnap` callback to remove the icon from the
pool, since that would prevent an immediately subsequent `doLinkSnap`
from finding the icon in the directory.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>