o/snapstate: clarify that linkSnapIcon returns an error if icon already linked
It may be useful to revise this behavior in the future, to allow
linkSnapIcon to be idempotent. Otherwise, if a snap icon is already
linked to the icons directory, but a new icon is downloaded, attempting
to link that new icon will fail and the old icon will still be linked.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>