tests/assertions: fix recovery-system-reboot:factory_reset test
recovery-system-reboot:factory_reset creates a recovery system and
does a factory reset using it, also using a validation-set assertion
that pins some snaps, including pc-kernel 1869. That revision has a
kernel signed with a Canonical key. However, when a refresh is
triggered after the factory reset, the system refreshes to a pc-kernel
from 24/edge as this is the channel specified in the model used in the
test. This kernel is signed with a test key. This makes the refresh
fail as Secure Boot is enabled. This seems to be random as the
re-refresh some times does not refresh all the snaps pending.

This change changes the model assertion to use pc-kernel from 24/beta so
we have a kernel with a good signature.

Fwiw, the original seed created by the test is fine because we
override the model channel to beta in nested_create_core_vm.