data/udev/rules.d: remove snapd auto-import rules 66-snapd-autoimport.rules not required for deb package (#14912)
* Remove udev snapd autoimport rule

This commit relates to LP#1966203, which reported the presence of syslog errors
associated with snapd auto-import that were not representative of a functional
issue on Ubuntu Desktop or Server.

A summary of the root cause is that systemd introduced a SystemCallFilter to
units/systemd-udevd.service.in in commit ee8f26180d which restricted the
syscalls available to udev. However snapd's 66-snapd-autoimport.rules attempts
to import snap assertions from connected block devices, which conflicts with the
filter.

As per comment #6 in the LP bug, the snapd team determined that this udev rules
file should not be triggered on desktop or server because the functionality is
intended only for Ubuntu Core. This commit removes the rule as outlined in
comment #13 in the LP bug, which expressed intent to remove it from the snapd
deb.

Related Bugs: LP#1966203, LP#1971955

* Fix: Remove 66 udev rule file from install template

* Fix: Remove 66 udev rule file from install template