o/snapstate: only discard store metadata if no other instances present
Since store metadata is not specific to any particular revision or
instance of a snap, it should not be removed if there are other
instances of the snap still present.

By making `hasOtherInstances` a parameter to `DiscardStoreMetadata`,
callers are encouraged to compute and check `hasOtherInstances`
beforehand, thus avoiding accidentally removing metadata which is
expected for another instance of the same snap.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>