cmd/snap: fix behavior of snap get with tty (#15037)
* cmd/snap: fix behavior of snap get with tty

The output of 'snap get' command changes depending on whether the
process is attached to a tty or not. However, instead of checking
whether stdout is attached to a tty, and thus produce a compact listing,
the code checked stdin. Thus redirecting the output did not produce a
desired change in the format.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/snap: group tty related variables in one place

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>