cmd/snap-bootstrap: mount base directly on /sysroot (#14537)
* cmd/snap-bootstrap: mount base directly on /sysroot

Mount the base directly on /sysroot in UC modes, while for hybrid we
mount /run/mnt/data instead of the /run/mnt/sysroot symlink that was
pointing to the former.

* core-initrd: adapt to changes now that the base mounts to /sysroot

With the snap-bootstrap changes, the base/rootfs is mounted directly
to /sysroot. With this, we do not need the symlinks created by
detect-*-sysroot.service units. However, we need to keep the static
sysroot mount unit so dependencies are still honored.

kernel-snap-generator and classic-mounts.service are not needed as for
24+ the kernel mount units are created by snap-bootstrap.

The units that need to detect if we are on core/hybrid check now for
the /run/mnt/data/system-data (if present, it is UC) instead of for a
mountpoint in /run/mnt/base.

* cmd/snap-bootstrap: restrict the change in base mount to UC24+

* tests/core/basic20plus: adapt to change in mount of sysroot

* cmd/snap-bootstrap: do not try to read unmounted base on installation

for UC24+.

* cmd/snap-bootstrap: mount base in recover mode

We need to mount the base so we can build the password related files.

* cmd/snap-bootstrap: restart initrd-root-fs.target

Restart initrd-root-fs.target so dependencies are recalculated after
the sysroot.mount unit is created.

* core-initrd: remove static sysroot.mount and unneeded dependencies

sysroot.mount is defined now dynamically by
snap-initramfs-mounts.service.