o/snapstate: move store metadata calls to within backend methods
Move `installStoreMetadata` call from the snapstate handler into the
backend `LinkSnap` method call, and move `uninstallStoreMetadata` into
`UnlinkSnap`. This reduces the complexity of the snapstate handlers.

Since `doDiscardSnap` still needs to discard store metadata if there are
no other revisions/instances of the snap present, and it is not
inherently associated with a particular install/uninstall operation,
there is an additional new metadata helper which, instead of taking a
`LinkContext`, simply takes a `hasOtherInstances` bool, and uses it to
construct a `LinkContext` and then call `uninstallStoreMetadata`.
Auxiliary store info is removed regardless of whether `UnlinkSnap` or
`doDiscardSnap` are the callers, but other store metadata (such as
cached snap icons) may behave differently, so this wrapper will enable
this distinction in the future.

One important note: auxinfo for a snap should not be removed unless
there are no other revisions of the snap present, because in case the
unlink snap is reverted, `undoUnlinkSnap` needs to read the auxinfo from
disk in order to populate `oldInfo`, so it can be re-linked.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>