interfaces: add support for snapped pipewire (#14453)
* audio-playback: Add support for snapped pipewire

A snapped pipewire requires being able to create the pulse
folder inside the global XDG_RUNTIME_DIR folder (not the snap
own folder). Also, it requires to be able to create both the
pipewire-[0-9]-manager and its lock file.

This patch adds that capability to the audio-playback slot;
thus, a snap offering that slot will be able to create the
corresponding global sockets.

* Add global tests

* fix extra tests

* Remove slot in non-core

* Remove the other

* Set install permissions

* Re-work the /etc/pulse access

* Fix tests

* Remove unneeded access to manager socket

* The snap must be able to be used in Core

* Fix pipewire test

* Fixed test and added changes requested by James

* Fixes requested by James

* Fix copyright date