o/snapstate{,/backend}: split metadata removal helper into uninstall and discard
When a snap is being unlinked or a link is being undone, the snapstate
handler should call `UnlinkStoreMetadata`, which takes a `LinkContext`
and checks whether there are other revisions of the snap present.

When the final revision of a snap is being discarded, the snapstate
handler should call `DiscardStoreMetadata`, which takes a
`hasOtherInstances bool` and assumes that there are no other revisions
of that instance of the snap present.

At the moment, the latter simply calls the former, but in the future,
other store metadata which should be handled differently during unlink
and discard can be handled appropriately.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>