core-initrd,snap-bootstrap: look at env to mount directly to /sysroot
Previously we looked at the model and mounted base/rootfs directly if
it was Ubuntu 24+. Now, we check instead an environment variable that
is set if the initramfs is 24+.

This is done so now the kernels that contain the 24+ initramfs can be
booted in older Ubuntu releases with older bases. This situation is
not something we really support, as the systemd of the initramfs will
not match the one in the system, but it is something that could happen
while remodeling and it seems safer to allow this at least
temporarily.