store: record snap icon etag and use If-None-Match header (#15032)
* logger: add function to mock logger with debug enabled

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* logger: add test for function which mocks debug logger

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: record snap icon etag and use If-None-Match header

Store the etag for the snap icon (if etag exists) in the xattrs of the
icon file on disk. When attempting to update an existing icon, first try
to read the etag, and if it does exist, send it in the `If-None-Match`
header of the HTTP request to the server.

If the current store icon has the same etag as the existing icon on disk
(and the server behaves), the server should respond with 304 Not
Modified, and not send the whole icon contents in the response body. The
download icon function can then exit early, discarding its pending
changes to the filesystem.

If the etag from the server differs from that on disk (or one of them
lacks an etag), then the server should respond with a 200, and send the
whole icon file contents. The icon download function should save the
file with the new etag from the response, if it exists.

Xattr names must be namespaced, and the `user` namespace is the
appropriate one for arbitrary data like etag values. Thus, we use
`user.etag` as the xattr name, with the etag as the value.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: rename icon xattr user.etag to user.snapstore-etag

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* store: check snap icon etag size before writing it to xattrs

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: use resp.Header.Get() because response headers are already canonicalized

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: skip etag tests if xattrs unsupported

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: adjust etag handling, logging, and testing following review

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: decrease etag size limit to 256B and adjust error log message

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

* store: adjust etag xattr name and define as const

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>

---------

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>
Co-authored-by: Maciej Borzecki <maciej.borzecki@canonical.com>