gadget: make reinstalls honor existing sizes of partitions
When performing a re-install, snapd uses "size" from the gadget when
re-creating partitions that it just removed. However, in the case of
using min-size there is a possibility that the deleted partition size
was smaller but still compatible with the installed gadget. Use the
size of the deleted partition, as we want to respect the already
existing layout - otherewise we might destroy partitions that are not
re-created in the re-install and that could appear after the
partitions using min-size.