cmd/libsnap-confine-private/tool: switch identity only after forking a child
We had a recurring pattern of switching the identity, then calling a
tool, followed by restoring the old identity. Since the tool is called
in a forked child process anyway, there is no clear win to switch the
identity in the parent process, which then requires a followup restore
operation. Move switching of identity to the child process, this
simplifying the code path.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>