tests: add test for the posix-mq interface
The posix-mq interface was merged a while ago without any dedicated tests.
Historically there were a number of issues around mediation of mqueue and
having this sort of test provided earlier might have saved us some time later.

The test verifies runtime correctness of the read, write, create and delete
permissions specific to the posix-mq interface, using the test-snapd-posix-mq
snap. The snap has been issued with a snap declaration assertion to use the
posix-mq interface slots, which are super-privileged. All slots grant access to
the /test POSIX message queue, with the relevant permission, read, write,
create or delete.

The tests is skipped on Ubuntu Core 18 and earlier, and equivalent, due to a
bug (referenced). Technically the interface apparmor elements require
apparmor_parser version 4.0.1 or newer present in snapd snap. The interface
does not work with host-provided apparmor as snapd explicitly constrains it to
ABI 3.0 which does not support mqueue.

In addition also remove posix-mq plug of test-snapd-policy-app-consumer as that
is required by static checker:

  Dedicated test 'tests/main/interfaces-posix-mq' found for 'posix-mq'.
  Please remove 'posix-mq' from 'tests/lib/snaps/test-snapd-policy-app-consumer/meta/snap.yaml'.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>