cmd/libsnap-confine-private/tool: switch identity only after forking a child (#15100)
* cmd/libsnap-confine-private: tweak sc_identity to use bool fields

Switch sc_identity change_{uid,gid} to use bools, such that it's clearer
that those fields are not carrying actual uid/gid but only dictate the
action to be taken.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

* cmd/libsnap-confine-private/tool: switch identity only after forking a child

We had a recurring pattern of switching the identity, then calling a
tool, followed by restoring the old identity. Since the tool is called
in a forked child process anyway, there is no clear win to switch the
identity in the parent process, which then requires a followup restore
operation. Move switching of identity to the child process, this
simplifying the code path.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>

---------

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>