tests: run tests on master branch nightly (#15084)
* tests: run tests on master branch nightly

This change runs the tests using master branch nightly instead of after
a change is pushed to master.

We do that in order to reduce the amount of instances used in google.

Also the nightly job was renamed to nightly-static-analysis which is a
more representative name.

* Move from google and openstack to fundamental and not-fundamental

This is because there are several google-* backends and doing this we
simplify the workflow configuration

* Running nested tests from a new matrix definition in json file

* Update .github/workflows/nightly-spread.yaml

Co-authored-by: Katie May <63071677+maykathm@users.noreply.github.com>

* Adding nested systems to the read-system step output

* fix the SPREAD_USE_SNAPD_SNAP_URL used for arm64

---------

Co-authored-by: Katie May <63071677+maykathm@users.noreply.github.com>