core-initrd: make Azure kernels reboot upon panic
Add the "panic=60" kernel cmdline option for Azure kernels in order to
make them reboot after 60s upon a panic during boot. This will be
relevant for implementing a fallback mechanism in nullboot, allowing
rollbacks in the event of a faulty kernel upgrade. The 60s timeout is
chosen to allow easier collection of console logs during troubleshooting.

Signed-off-by: Vinicius Peixoto <vinicius.peixoto@canonical.com>