tests: fix snap-user-service-socket-activation with confined nc
AppArmor is undergoing a flurry of activity, with dozens of new
confinement profiles for previously unconfined applications. In Ubuntu
25.04 and newer, the nc program is now sandboxed and the sandbox does
not allow cap_dac_override. As such the test running as root, no longer
has the ability to traverse /home/test or /run/user/12345/ which
ordinarily lack the o+x permission.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>