tests/main/services-stop-mode-sigkill: move testing os sigterm stop-modes into separate snap
Split out testing of sigterm{,-all} stop modes into a separate snap. The
sigterm stop mode has a tendency to leak processes, which is a direct
effect of the requested KillMode=process and so requires additional
cleanup. However, the original snap test-snapd-service is used in a
number of tests, so in order to avoid additional changes for cleanup
needing to be added to each of those tests, the patchs splits out
relevant parts of the original snap into a new one which is used only in
a specific test targeting the stop-mode feature.

Signed-off-by: Maciej Borzecki <maciej.borzecki@canonical.com>