o/confdbstate: add helper to schedule load confdb hook/tasks (#15110)
* o/confdbstate: change id under which tx carrying task is set

The loading change will reuse GetStoredTransaction to obtain the
transaction stored or linked by a task. However, the confdb loading
doesn't commit anything so it doesn't make sense to store the task ID
under "commit-task". This changes that to "tx-task".

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* o/confdbstate: refactor gathering of custodian names

Refactors the accumulating of custodian snap names into the same
function that looks for relevant plugs so we don't need to duplicate it
in when we implement something similar for the loading confdb change.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

* o/confdbstate: schedule hooks/tasks to load confdbs

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>

---------

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>